notifications/requirements.txt
django==4.2
celery
sendgrid
python-telegram-bot
requests
redis